3.2 BRANCH: Build 2891.2943: Backport of r1939, r1940, r1941, r1942

+++ branches/3.2/tunnelblick/UKKQueue/UKFNSubscribeFileWatcher.m
@@ -197,5 +197,5 @@ void    UKFileSubscriptionProc( FNMessage message, OptionBits flags, void *refco
     if( message == kFNDirectoryModifiedMessage )    // No others exist as of 10.4
         [obj sendDelegateMessage: message forSubscription: subscription];
     else
-        NSLog( @"UKFileSubscriptionProc: Unknown message %d", message );
+        NSLog( @"UKFileSubscriptionProc: Unknown message %lu", (unsigned long) message );
 }
